+++ RELEASENOTES.md
@@ -54,6 +54,9 @@
 *   IMA extension
     *   Remove player listener of the `ImaServerSideAdInsertionMediaSource` on
         the application thread to avoid threading issues.
+    *   Add a property `focusSkipButtonWhenAvailable` to the
+        `ImaServerSideAdInsertionMediaSource.AdsLoader.Builder` to request
+        focusing the skip button on TV devices and set it to true by default.
     *   Bump IMA SDK version to 3.29.0.
 
 ### 1.0.0-beta03 (2022-11-22)